Merged PR 21: finally a successful deployment of the bot to azure
Currently there are still some tweaky things that have to happen to the Keys and App Registrations both here (AzDO, with pipeline variables) and Azure (with the App Registration) that I want to address moving forward.  But this gets me un-blocked to move on to both front-end rework, and Conversation Analyst work/development.